Add translate-attr Example
I did not see an example of using `translate-attr-*` in the guide. I started writing my own directive and was using the translate directive as a reference when I saw that the functionality did exist!

The guide has been updated to alert users that it is possible to translate attributes like `alt` on HTML attributes.